+++ src/string.rs
@@ -4,12 +4,13 @@ use crate::{
     alloc::NSTDAllocError,
     core::{
         def::{NSTDByte, NSTDErrorCode},
-        optional::{NSTDOptional, NSTDOptionalUnichar},
+        optional::NSTDOptional,
         slice::{nstd_core_slice_new, NSTDSlice},
         str::{
             nstd_core_str_as_bytes, nstd_core_str_from_bytes_unchecked, nstd_core_str_len,
             nstd_core_str_mut_from_bytes_unchecked, NSTDStr, NSTDStrMut,
         },
+        unichar::{NSTDOptionalUnichar, NSTDUnichar},
     },
     vec::{
         nstd_vec_as_ptr, nstd_vec_as_slice, nstd_vec_as_slice_mut, nstd_vec_cap, nstd_vec_clear,
@@ -17,7 +18,7 @@ use crate::{
         nstd_vec_new_with_cap, nstd_vec_truncate, NSTDVec,
     },
     NSTDFloat32, NSTDFloat64, NSTDInt, NSTDInt16, NSTDInt32, NSTDInt64, NSTDInt8, NSTDUInt,
-    NSTDUInt16, NSTDUInt32, NSTDUInt64, NSTDUInt8, NSTDUnichar,
+    NSTDUInt16, NSTDUInt32, NSTDUInt64, NSTDUInt8,
 };
 use alloc::string::ToString;
 
@@ -357,15 +358,13 @@ pub extern "C" fn nstd_string_cap(string: &NSTDString) -> NSTDUInt {
 /// ```
 #[cfg_attr(feature = "clib", no_mangle)]
 pub extern "C" fn nstd_string_push(string: &mut NSTDString, chr: NSTDUnichar) -> NSTDErrorCode {
-    if let Some(chr) = char::from_u32(chr) {
-        let mut buf = [0; 4];
-        chr.encode_utf8(&mut buf);
-        let buf = nstd_core_slice_new(buf.as_ptr().cast(), 1, chr.len_utf8());
-        // SAFETY: `buf`'s data is stored on the stack.
-        let errc = unsafe { nstd_vec_extend(&mut string.bytes, &buf) };
-        return (errc != NSTDAllocError::NSTD_ALLOC_ERROR_NONE).into();
-    }
-    1
+    let chr = char::from(chr);
+    let mut buf = [0; 4];
+    chr.encode_utf8(&mut buf);
+    let buf = nstd_core_slice_new(buf.as_ptr().cast(), 1, chr.len_utf8());
+    // SAFETY: `buf`'s data is stored on the stack.
+    let errc = unsafe { nstd_vec_extend(&mut string.bytes, &buf) };
+    (errc != NSTDAllocError::NSTD_ALLOC_ERROR_NONE).into()
 }
 
 /// Appends a string slice to the end of a string.
@@ -448,7 +447,7 @@ pub extern "C" fn nstd_string_pop(string: &mut NSTDString) -> NSTDOptionalUnicha
     if let Some(chr) = str.chars().last() {
         let len = nstd_vec_len(&string.bytes) - chr.len_utf8();
         nstd_vec_truncate(&mut string.bytes, len);
-        return NSTDOptional::Some(chr as NSTDUnichar);
+        return NSTDOptional::Some(chr.into());
     }
     NSTDOptional::None
 }